py-ai/requirements.txt
langchain
openai
python-dotenv
pydantic
langchain_community
fastapi
uvicorn
bs4
chromadb
tiktoken